chore: Setup basic project structure and config (#1)
* chore: Add initial configuration files

- Add .clang-format for code formatting.
- Add .clang-tidy for code linting.
- Add .gitattributes for normalizing line endings.
- Add .gitignore to ignore unnecessary files.

* chore: Add submodules for cocos2d-x and googletest

- Add .gitmodules file to track submodules.
- Add cocos2d-x, googletest as submodules to decouple them from the
  repository.

* chore: Add licenses of third-party libraries

* docs: Add blank design document

* chore: Add custom CMakeLists for better control

Create custom `CMakeLists.txt` instead of using the one generated by
the `cocos new` command to provide greater control over the build
process and file structure. This approach eliminates the need to copy
the entire cocos2d-x framework source files into the project, resulting
in a cleaner and more organized project structure.

* feat: Add sample application and scene

Add a sample application and scene as a demonstration and sandbox for
testing and experimenting with the cocos2d-x framework. This also
defines the initial file structure for the project.

* test: Add Google Test for sample application

* feat: Add src/main.cpp to run sample application

Create src/main.cpp to run the sample application, testing the build
system and ensuring the project and environment are set up correctly.

* chore: Add fonts for sample application

Add `DFHannotateW5-A.ttf` and `PlayfairDisplay-Regular.ttf` to the
project. These fonts are used in the sample application to validate font
loading and ensure proper resource path resolution.

* chore: Update CMakeLists.txt for Windows support

- Add conditional compiler flags for MSVC to ensure compatibility.
- Add a conditional block to copy dynamic-link libraries (DLLs) on
  Windows.

* chore: Fix GTest integration in CMakeLists.txt

- Correct the logic to find and use the installed GoogleTest library.

* fix: Correct GoogleTest integration logic

- Fixed the condition to correctly check if GTest is found.

* chore: Optimize CMakeLists.txt for compatibility

- Lower minimum required CMake version to 3.15 for better compatibility
  with older versions.
- Introduce a BUILD_TESTS option to conditionally build tests.
  - This option is added because tests rely on the Google Test
    framework, which is downloaded and built by CMake fetch content.
    Network restrictions can sometimes prevent this, causing the CMake
    configuration to fail.

* chore: Add CMakeSettings for VS 2022 x86 builds

The cocos2d-x project can compile only with x86 architecture on MSVC.
The CMakeSettings.json file is added to the project to set the
architecture to x86 and use the Ninja generator for faster builds.

* chore: Update .gitignore

Ignore output directories of Visual Studio 2022.

* fix: Remove experimental namespace for AudioEngine

The experimental namespace for the AudioEngine class has been removed as
it is no longer experimental in Windows.

* chore: Cache fetched files in .cmakecache

Add FETCHCONTENT_BASE_DIR to CMakeLists.txt to cache downloaded files
in .cmakecache directory. This change avoids downloading the files
again, improving build efficiency.

* chore: Add .editorconfig to enforce code style

- Use 4 spaces for indentation.
- Use LF for line endings (Unix-style).
- Insert a final newline at the end of the file.
- Trim any trailing whitespace.

* fix: correct Makefile format with backslash

In the Makefile, every line should not start with space, otherwise it
will cause an error when running make command. To avoid this issue, we
add a backslash in CMAKE_CXX_FLAGS to separate the flags.

* refactor: Normalize line endings and format files

* chore: delete config for line endings

We don't use .editorconfig to normalize line endings because it can
cause mixed line endings in the same file. Instead, we use the
.gitattributes file to normalize line endings.

* chore: Save files before normalizing line endings