Fix: Use multi thread in NetworkClient for asio
Currently, the NetworkClient is using a single thread to handle the
networking. This is not optimal, as it can block the main thread if no
messages are being received. This commit changes the NetworkClient to
use a separate thread for io_context.run(), which solves the blocking
main thread issue.