fix: Store sent/received messages in the class
Currently, the messages are not stored in the class, which causes
`cannot dereference string iterator because the iterator was
invalidated` error in Windows with msvc. This commit fixes the issue
partially by storing the sent/received messages in the class.

After this commit, the messages are stored in the class and reduce the
possibility of the error from 100% to about 30%.