chore: Use AddressSanitizer for debugging
The error `cannot dereference string iterator because the iterator was
invalidated` happens with a possibility of about 30%, to debug this
issue, we try to use AddressSanitizer to find the root cause.

After running the program with AddressSanitizer, the AddressSanitizer
does not report any error, which means the issue is not caused by
memory corruption.